requirements.txt
appdirs==1.4.4
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
pyparsing==2.4.7
python-Levenshtein==0.12.0
PyYAML==5.3.1
six==1.15.0
Werkzeug==1.0.1
numpy==1.18.5
